refactor: update MCP entry point and remove obsolete main module
- Changed the entry point for the `chift-mcp-server` script to point to `chift_mcp.run_local:main`.
- Removed the `__main__.py` file as it is no longer needed, streamlining the codebase.
